+++ components/pages/blog/blogs/hostingANextjsServerlessStaticSiteUsingAWSCDK/controller.tsx
@@ -1,5 +1,5 @@
 import Index from './index'
-import { blogs } from '../controller'
+import { blogs } from '../../controller'
 
 const Component = (): JSX.Element => <Index blogMetaData={blogs.hostingANextjsServerlessStaticSiteUsingAWSCDK} />
