+++ pages/blog/hosting-a-nextjs-serverless-static-site-using-aws-cdk.tsx
@@ -1,4 +1,4 @@
-import Blog from '../../components/pages/blog/hostingANextjsServerlessStaticSiteUsingAWSCDK/controller'
+import Blog from '../../components/pages/blog/blogs/hostingANextjsServerlessStaticSiteUsingAWSCDK/controller'
 
 const Page = (): JSX.Element => <Blog />
